Add new classes to organise code base
