+++ src/main/java/data/tasks.txt
@@ -1,3 +1,4 @@
 T | 0 | borrow book
-D | 0 | return book | Sunday
-E | 0 | project meeting | Mon 2-4pm
+T | 0 | test
+T | 0 | test1
+T | 0 | test2
